Calculate speed from moved distance and time since last event. TG-64 (#65)
* Calculate speed from moved distance and time

* Store config keys in enum instead of properties file

* Add config entry for speed check delta

* Update Train tests

* Fix format string in InterpolatableState.Builder

+++ src/main/java/com/github/bachelorpraktikum/dbvisualization/view/detail/TrainDetail.java
@@ -38,7 +38,7 @@ public class TrainDetail extends ElementDetailBase<Train> {
         return getElement().getState(timeProperty().get());
     }
 
-    int getSpeed() {
+    double getSpeed() {
         return getState().getSpeed();
     }
